+++ pyinaturalist/response_format.py
@@ -39,6 +39,7 @@ def as_geojson_feature_collection(
         results: List of results from API response
         properties: Whitelist of specific properties to include
     """
+    results = [flatten_nested_params(obs) for obs in results]
     return {
         'type': 'FeatureCollection',
         'features': [as_geojson_feature(record, properties) for record in results],
@@ -228,7 +229,7 @@ def try_float(value: Any) -> Optional[float]:
         return None
 
 
-# Formatting functions
+# Formatting Functions
 # --------------------
 
 
@@ -250,7 +251,6 @@ def flatten_nested_params(observation: ResponseObject) -> ResponseObject:
     return observation
 
 
-# TODO: Is any error handling necessary here? Responses seem fairly consistent and predictable.
 def format_histogram(response: JsonResponse) -> HistogramResponse:
     """Format a response containing time series data into a single ``{date: value}`` dict"""
     # The inner result object's key will be the name of the interval requested
@@ -262,33 +262,3 @@ def format_histogram(response: JsonResponse) -> HistogramResponse:
         return {int(k): v for k, v in histogram.items()}
     else:
         return {parse_date(k): v for k, v in histogram.items()}
-
-
-def format_observation(observation: ResponseObject) -> str:
-    """Make a condensed summary from basic observation details: what, who, when, where"""
-    taxon_str = format_taxon(observation.get('taxon') or {})
-    location = observation.get('place_guess') or observation.get('location')
-    return (
-        f"[{observation['id']}] {taxon_str} "
-        f"observed by {observation['user']['login']} "
-        f"on {observation['observed_on']} "
-        f'at {location}'
-    )
-
-
-def format_taxon(taxon: ResponseObject, align: bool = False) -> str:
-    """Format a taxon result into a single string containing taxon ID, rank, and name
-    (including common name, if available).
-    """
-    if not taxon:
-        return 'unknown taxon'
-    common_name = taxon.get('preferred_common_name')
-    name = f"{taxon['name']}" + (f' ({common_name})' if common_name else '')
-    rank = taxon['rank'].title()
-
-    # Visually align taxon IDs (< 7 chars) and ranks (< 11 chars)
-    if align:
-        # return '{:>8}: {:>12} {}'.format(taxon['id'], rank, name)
-        return f"{taxon['id']:>8}: {rank:>12} {name}"
-    else:
-        return f'{rank}: {name}'